chore(ton): split adapter into constants, utils, parser modules
TonChainAdapter.ts carried every type, constant and pure helper
alongside the adapter class (~1360 lines). Constants move to
constants.ts, address/hash helpers to utils.ts, transfer parsing
(parseTonTx, buildJettonTransfers, buildTraceTransfers) to parser.ts,
and the remaining local types to types.ts - the adapter file keeps only
the class. All modules re-exported from the ton index; no behavior
changes (live parse output verified identical).
