+++ packages/chain-adapters/src/ton/index.ts
@@ -1,3 +1,6 @@
 export { ChainAdapter } from './TonChainAdapter'
 
+export * from './constants'
+export * from './parser'
 export * from './types'
+export * from './utils'
